Revert "Rc1 patch and Tests"
