--- .github/FUNDING.yml
@@ -1,12 +1,0 @@
-# These are supported funding model platforms
-
-# github: [ZauJulio, wanessabezerra, RobertoCostaTupinamba] # Replace with up to 4 GitHub Sponsors-enabled usernames e.g., [user1, user2]
-patreon: ZauJulio # Replace with a single Patreon username
-open_collective: # Replace with a single Open Collective username
-ko_fi: # Replace with a single Ko-fi username
-tidelift: # Replace with a single Tidelift platform-name/package-name e.g., npm/babel
-community_bridge: # Replace with a single Community Bridge project-name e.g., cloud-foundry
-liberapay: # Replace with a single Liberapay username
-issuehunt: # Replace with a single IssueHunt username
-otechie: # Replace with a single Otechie username
-custom: # Replace with up to 4 custom sponsorship URLs e.g., ['link1', 'link2']
