--- taskiano/.babelrc
@@ -1,5 +1,0 @@
-{
-  "presets": [
-    "next/babel"
-  ]
-}
