+++ taskiano/.eslintrc
@@ -1,3 +1,3 @@
 {
-  "extends": ["next/babel"]
+  "extends": ["next", "next/core-web-vitals"]
 }
