--- taskiano/__tests__/lib/controllers/FireController.ts
@@ -1,3 +1,0 @@
-import FireController from "../../mocks/FireController.mock";
-
-export default FireController;
