--- taskiano/__tests__/lib/controllers/index.ts
@@ -1,6 +1,0 @@
-import UserController from "./UserController";
-import TaskController from "./TaskController";
-import ProjectController from "./ProjectController";
-import HistoryController from "./HistoryController";
-
-export { UserController, TaskController, ProjectController, HistoryController };
