--- taskiano/__tests__/lib/index.ts
@@ -1,8 +1,0 @@
-import {
-  UserController,
-  ProjectController,
-  TaskController,
-  HistoryController,
-} from "./controllers";
-
-export { UserController, ProjectController, TaskController, HistoryController };
