--- taskiano/__tests__/mocks/data/historys.ts
@@ -1,20 +1,0 @@
-import { IHistory } from "../../../types";
-
-export let historys: IHistory[] = [
-  {
-    id: "6c325307-8ee1-4055-bb38-a4704a272ac2",
-    score: 666,
-    updated_at: new Date(),
-    weekdayTaskCount: {
-      mon: 0,
-      tue: 0,
-      wed: 0,
-      thu: 0,
-      fri: 0,
-      sat: 0,
-      sun: 0,
-    },
-    lastTaskNumber: 4,
-    userId: "7fd18545-8054-45f5-9d76-9addd179d13b",
-  },
-];
