--- taskiano/__tests__/mocks/data/users.ts
@@ -1,12 +1,0 @@
-import { IUser } from "../../../types";
-
-export let users: IUser[] = [
-  {
-    id: "7fd18545-8054-45f5-9d76-9addd179d13b",
-    username: "taciano",
-    name: "Taciano",
-    avatar: "https://github.com/tacianosilva.png",
-    birthday: new Date("1902-12-23"),
-    email: "taciano.silva@ufrn.br",
-  },
-];
