--- taskiano/__tests__/types/Collections.ts
@@ -1,5 +1,0 @@
-interface Collections {
-  [key: string]: any[];
-}
-
-export default Collections;
