+++ taskiano/contexts/AuthContext/index.tsx
@@ -2,7 +2,7 @@ import { ReactNode, useEffect, useState } from "react";
 import type { NextRouter } from "next/router";
 
 import { AuthContext } from "./Provider";
-import { UserController } from "../../lib";
+import { UserController } from "../../db";
 import useFirebaseAuth from "../../hooks/useFirebaseAuth";
 import { ToastDisconnected, ToastTrySignInAgain } from "../../utils/toasts";
 import type { IUser } from "../../types";
