+++ taskiano/db/schemas/UserSchema.ts
@@ -2,6 +2,7 @@ import * as Yup from "yup";
 
 const UserSchema = Yup.object().shape({
   id: Yup.string().required(),
+  countCreatedTasks: Yup.number().required(),
   username: Yup.string().required(),
   avatar: Yup.string().required(),
   birthday: Yup.date(),
