--- taskiano/env.local.example
@@ -1,14 +1,0 @@
-# Use secure-env package to set Firebase encrypted environment variables in .env.enc
-# https://www.npmjs.com/package/secure-env
-
-API_KEY=xxxxxxxxxxxxxxxxxxxxxxxxxxxx
-AUTH_DOMAIN=xxxxxxxxxxxxxxxxxxxxxxxxxxxx
-PROJECT_ID=xxxxxxxxxxxxxxxxxxxxxxxxxxxx
-STORAGE_BUCKET=xxxxxxxxxxxxxxxxxxxxxxxxxxxx
-MESSAGING_SENDERID=xxxxxxxxxxxxxxxxxxxxxxxxxxxx
-APP_ID=xxxxxxxxxxxxxxxxxxxxxxxxxxxx
-MEASUREMENT_ID=xxxxxxxxxxxxxxxxxxxxxxxxxxxx
-
-# And save a file .env.local with:
-
-SECURE_ENV=xxxxxxxxxxxxxxxxxxxxxxxx
