+++ taskiano/hooks/useFirebaseAuth.tsx
@@ -10,7 +10,7 @@ import {
   User as IFirebaseUser,
 } from "firebase/auth";
 
-import { UserController } from "../lib";
+import { UserController } from "../db";
 import { auth } from "../services/Firebase";
 
 import {
