--- taskiano/jest.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  setupFilesAfterEnv: ["<rootDir>/jest.setup.js"],
-  testPathIgnorePatterns: ["<rootDir>/.next/", "<rootDir>/node_modules/"],
-  moduleNameMapper: {
-    "\\.(scss|sass|css)$": "identity-obj-proxy",
-  },
-  testMatch: ['**/?(*.)+(spec|test).ts(x)?']
-};
