+++ taskiano/package.json
@@ -1,20 +1,17 @@
 {
   "name": "taskiano",
-  "version": "1.1.0",
+  "version": "1.0.0",
   "license": "MIT",
   "private": true,
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage"
+    "lint": "next lint"
   },
   "dependencies": {
     "firebase": "9.6.0",
-    "next": "12.0.5",
+    "next": "12",
     "pathseg": "^1.2.1",
     "react": "^17.0.2",
     "react-countdown-circle-timer": "^2.5.3",
@@ -32,27 +29,14 @@
     "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@babel/core": "^7.16.5",
-    "@babel/preset-env": "^7.16.5",
-    "@babel/preset-typescript": "^7.16.5",
-    "@testing-library/dom": "^8.11.1",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.0.3",
     "@types/react": "17.0.37",
     "@types/sass": "^1.16.1",
-    "@types/uuid": "^8.3.3",
     "@types/yup": "^0.29.13",
-    "@typescript-eslint/eslint-plugin": "^5.7.0",
-    "babel-jest": "^27.4.5",
     "eslint": "<8.0.0",
     "eslint-config-next": "12.0.4",
-    "jest": "^27.4.5",
     "sass": "^1.36.0",
     "secure-env": "^1.2.0",
-    "typescript": "^4.4.4",
-    "uuid": "^8.3.2"
+    "typescript": "^4.4.4"
   },
   "peerDependencies": {
     "@babel/core": "^7.16.0",
