+++ taskiano/tsconfig.json
@@ -17,13 +17,12 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "preserve",
-    "incremental": true,
+    "incremental": true
   },
   "include": [
     "next-env.d.ts",
     "**/*.ts",
-    "**/*.tsx",
-    "pages/_document.js"
+    "**/*.tsx"
   ],
   "exclude": [
     "node_modules"
